cmake: let USE_ICONV be optional on macOS
Instead of forcing iconv support on macOS (by forcing `USE_ICONV`
on), set up a (cached) variable for `USE_ICONV` that defaults to
`AUTO`.  On macOS, `AUTO` will be `ON`, otherwise `OFF`, but users
can force `ON` or `OFF` to override the autodetect behavior.

Although macOS includes iconv by default, some macOS users may have a
deficient installation for some reason and they should be provided a
workaround to use libgit2 even in this situation.